+++ test/array_test.dart
@@ -38,5 +38,43 @@ void main() {
       expect(arrWithNumber(0, 5), [0, 1, 2, 3, 4, 5]);
       expect(arrWithNumber(1, 1), [1]);
     });
+
+    test('arrUnique', () {
+      List<List<int>> big2dArray = [
+        [10, 20, 30, 40, 50],
+        [1, 2, 3, 4, 5],
+        [6, 7, 8, 9, 0]
+      ];
+
+      funcTimes(150000, () => big2dArray.add([1, 1, 1, 1, 1]));
+      funcTimes(150000, () => big2dArray.add([2, 2, 2, 2, 2]));
+      funcTimes(150000, () => big2dArray.add([3, 3, 3, 3, 3]));
+
+      expect(arrUnique(big2dArray), [
+        [10, 20, 30, 40, 50],
+        [1, 2, 3, 4, 5],
+        [6, 7, 8, 9, 0],
+        [1, 1, 1, 1, 1],
+        [2, 2, 2, 2, 2],
+        [3, 3, 3, 3, 3]
+      ]);
+      expect(arrUnique([1, 1, 2, 2, 2, 2, 3]), [1, 2, 3]);
+      expect(arrUnique(['1', '2', '3', '3', '4']), ['1', '2', '3', '4']);
+      expect(arrUnique([1, '1', 1, 'a', 2, 'b']), [1, '1', 'a', 2, 'b']);
+      expect(
+          arrUnique([
+            [1, 2],
+            [1, 2],
+            [2, 3],
+            [2, 3],
+            [2, 3],
+            [2, 4]
+          ]),
+          [
+            [1, 2],
+            [2, 3],
+            [2, 4]
+          ]);
+    });
   });
 }